兼容1.1.1和1.2.0 token不一致问题

+++ miner-pool-config/src/main/java/com/mei/hui/config/HttpRequestUtil.java
@@ -27,7 +27,7 @@ public class HttpRequestUtil {
         Claims claims = parseToken();
         Object currency = claims.get(SystemConstants.CURRENCYID);
         if(currency != null){
-            Long currencyId = (Long) currency;
+            Integer currencyId = (Integer) currency;
             return Long.valueOf(currencyId);
         }
         return null;
